src/requirements.txt
django
djangorestframework
celery
numpy
scipy
networkx
scikit-fuzzy
matplotlib
redis
mysqlclient
mariadb
